chapter_001/src/main/java/ru/job4j/calculator/Calculator.java
package ru.job4j.calculator;

public class Calculator {

    private double result;

    public void add(double first, double second) {
        this.result = first + second;
    }

    public void subtraction(double first, double second){
        this.result = first - second;
    }

    public void multiply(double first, double second){
        this.result = first*second;
    }

    public void division(double first, double second){
        if(second !=0){
            this.result = first/second;
        }
    }

    public double getResult() {
        return this.result;
    }
}
